profiler: make profiler optional with GGML_GRAPH_PROFILER
